requirements.txt
langchain
streamlit
python-dotenv
unstructured
tiktoken
faiss-cpu
libmagic
python-magic
openai
